More Precommit fixes
