+++ library/pyproject.toml
@@ -215,6 +215,7 @@ notice-rgx = '# Copyright \(C\) (\d{4}(-\d{4})?) Intel Corporation\r?\n# SPDX-Li
 "tests/**/*.py" = ["PLR6301"]  # Ignore no-self-use for test methods
 
 
+
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # MYPY CONFIGURATION.                                                         #
 [tool.mypy]
