+++ library/src/getiprompt/models/foundation/efficientsam3/utils/__init__.py
@@ -1,0 +1,11 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+# TODO: Consider moving SAM1/SAM2 utilities to a shared location
+# if SAM3 also needs these transforms in the future.
+
+"""SAM1/SAM2 style utilities for EfficientSAM3."""
+
+from .sam1_utils import SAM2Transforms
+
+__all__ = ["SAM2Transforms"]
