--- library/src/getiprompt/models/foundation/sam3/sam/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved
-from .mask_decoder import MaskDecoder as MaskDecoder
-from .prompt_encoder import PromptEncoder as PromptEncoder
-from .transformer import TwoWayTransformer as TwoWayTransformer
